requirements.txt
anyio==3.6.2
APScheduler==3.6.3
bottle==0.12.23
cachetools==4.2.2
certifi==2022.12.7
charset-normalizer==3.0.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
pycodestyle==2.10.0
python-dotenv==0.21.1
python-telegram-bot==20.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
rfc3986==1.5.0
six==1.13.0
sniffio==1.3.0
tomli==2.0.1
tornado==6.1
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
